moved cards into card-list
