+++ utils/fetch-data.js
@@ -3,6 +3,7 @@ const getData = async (url, options) => {
     const res = await fetch(url)
     const json = await res.json()
     options.response = json
+    options.recipesList = options.response.map(card => card.name) // creates an array with recipes
   } catch (error) {
     options.error = error
   } finally {
